feat: 채팅/포인트/회원가입시 본인 확인메일

+++ todaytest-master/src/main/java/com/hobby/domain/UserVO.java
@@ -44,8 +44,8 @@ public class UserVO {
     private String usrChangePwdDate;
     // 회원 권한 리스트
     private List<AuthVO> authList;
-    // 포인트
-    private int usrPoint;
+    // 사용자 보유 포인트
+    private Long usrPoint;
     
     //userdetail 테이블
     // 관심지역1
